+++ src/shared/helpers/removeItem.ts
@@ -1,0 +1,3 @@
+export const removeItem = (): void => {
+  localStorage.removeItem("currentUser");
+};
